requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==8.11.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
